[RTE] InputBox and SendBox (#4007)
* Update inputBox

* OnChange

* Split components

* InputBox and SendBox

* Placeholder

* Add styles

* Styles v8

* Update api

* Beta flag

* placeholder

* strings

* update storybook

* Update sendbox

* Refactor InputBox

* Clean up refactor

* Remove children

* Change files

* Duplicate change files for beta release

* Clean up

* Address PR comments

* Use sendbox style in RTESendbox

* Fix lint
