+++ change-beta/@azure-communication-react-3098c052-5689-4b78-af33-4148fa1a08e2.json
@@ -1,0 +1,9 @@
+{
+  "type": "none",
+  "area": "feature",
+  "workstream": "RTE",
+  "comment": "InputBox and SendBox",
+  "packageName": "@azure/communication-react",
+  "email": "73612854+palatter@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
