+++ packages/react-components/src/components/ChatMessage/ChatMessageComponentAsEditBox.tsx
@@ -8,7 +8,8 @@ import { _formatString } from '@internal/acs-ui-common';
 import { useTheme } from '../../theming/FluentThemeProvider';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { editBoxStyle, inputBoxIcon, editingButtonStyle, editBoxStyleSet } from '../styles/EditBox.styles';
-import { InputBoxButton, InputBoxComponent } from '../InputBoxComponent';
+import { InputBoxComponent } from '../InputBoxComponent';
+import { InputBoxButton } from '../InputBoxButton';
 import { MessageThreadStrings } from '../MessageThread';
 import { useChatMyMessageStyles } from '../styles/MessageThread.styles';
 import { ChatMessage } from '../../types';
