+++ packages/react-components/src/components/InputBoxComponent.tsx
@@ -1,35 +1,19 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import React, { useState, ReactNode, FormEvent, useCallback } from 'react';
+import React, { ReactNode, FormEvent, useCallback } from 'react';
 
-import {
-  Stack,
-  TextField,
-  mergeStyles,
-  IStyle,
-  ITextField,
-  concatStyleSets,
-  IconButton,
-  TooltipHost,
-  ICalloutContentStyles,
-  ITextFieldProps
-} from '@fluentui/react';
+import { Stack, TextField, mergeStyles, IStyle, ITextField, concatStyleSets, ITextFieldProps } from '@fluentui/react';
 import { BaseCustomStyles } from '../types';
 import { isEnterKeyEventFromCompositionSession } from './utils';
 
 import {
   inputBoxStyle,
   inputBoxWrapperStyle,
-  inputButtonStyle,
   textFieldStyle,
-  textContainerStyle,
-  inputButtonTooltipStyle,
-  iconWrapperStyle
+  textContainerStyle
 } from './styles/InputBoxComponent.style';
 
-import { isDarkThemed } from '../theming/themeUtils';
-import { useTheme } from '../theming';
 /* @conditional-compile-remove(mention) */
 import { MentionLookupOptions } from './MentionPopover';
 /* @conditional-compile-remove(mention) */
@@ -178,56 +162,5 @@ export const InputBoxComponent = (props: InputBoxComponentProps): JSX.Element =>
     <Stack className={mergedRootStyle}>
       <div className={mergedTextContainerStyle}>{renderTextField()}</div>
     </Stack>
-  );
-};
-
-/**
- * Props for displaying a send button besides the text input area.
- *
- * @private
- */
-export type InputBoxButtonProps = {
-  onRenderIcon: (isHover: boolean) => JSX.Element;
-  onClick: (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => void;
-  className?: string;
-  id?: string;
-  ariaLabel?: string;
-  tooltipContent?: string;
-};
-
-/**
- * @private
- */
-export const InputBoxButton = (props: InputBoxButtonProps): JSX.Element => {
-  const { onRenderIcon, onClick, ariaLabel, className, id, tooltipContent } = props;
-  const [isHover, setIsHover] = useState(false);
-  const mergedButtonStyle = mergeStyles(inputButtonStyle, className);
-
-  const theme = useTheme();
-  const calloutStyle: Partial<ICalloutContentStyles> = { root: { padding: 0 }, calloutMain: { padding: '0.5rem' } };
-
-  // Place callout with no gap between it and the button.
-  const calloutProps = {
-    gapSpace: 0,
-    styles: calloutStyle,
-    backgroundColor: isDarkThemed(theme) ? theme.palette.neutralLighter : ''
-  };
-  return (
-    <TooltipHost hostClassName={inputButtonTooltipStyle} content={tooltipContent} calloutProps={{ ...calloutProps }}>
-      <IconButton
-        className={mergedButtonStyle}
-        ariaLabel={ariaLabel}
-        onClick={onClick}
-        id={id}
-        onMouseEnter={() => {
-          setIsHover(true);
-        }}
-        onMouseLeave={() => {
-          setIsHover(false);
-        }}
-        // VoiceOver fix: Avoid icon from stealing focus when IconButton is double-tapped to send message by wrapping with Stack with pointerEvents style to none
-        onRenderIcon={() => <Stack className={iconWrapperStyle}>{onRenderIcon(isHover)}</Stack>}
-      />
-    </TooltipHost>
   );
 };
