+++ packages/react-components/src/components/RTE/RTEInputBoxComponent.tsx
@@ -1,0 +1,32 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import React from 'react';
+import { BaseCustomStyles } from '../../types';
+import { RichTextEditor } from './RichTextEditor';
+
+/**
+ * @private
+ */
+export interface RTEInputBoxComponentStylesProps extends BaseCustomStyles {}
+
+/**
+ * @private
+ */
+export interface RTEInputBoxComponentProps {
+  placeholderText?: string;
+  content: string;
+}
+
+/**
+ * @private
+ */
+export const RTEInputBoxComponent = (props: RTEInputBoxComponentProps): JSX.Element => {
+  const { placeholderText, content } = props;
+
+  return (
+    <div>
+      <RichTextEditor content={content} placeholderText={placeholderText} onChange={() => {}} />
+    </div>
+  );
+};
