+++ packages/react-components/src/components/styles/InputBoxComponent.style.ts
@@ -75,39 +75,3 @@ export const textFieldStyle: IStyle = {
     borderRadius: '0.25rem'
   }
 };
-
-/**
- * @private
- */
-export const inputButtonStyle = mergeStyles({
-  color: 'grey',
-  margin: 'auto',
-  width: '1.0625rem',
-  height: '1.0625rem',
-  backgroundColor: 'transparent',
-  '&:hover': {
-    backgroundColor: 'transparent'
-  }
-});
-
-/**
- * @private
- */
-export const sameLineButtonsContainerStyle: IStyle = {
-  alignSelf: 'end'
-};
-
-/**
- * @private
- */
-export const inputButtonTooltipStyle = mergeStyles({
-  // The toolTip host container show be a flex box, so that alignItems: 'center' works for inside buttons
-  display: 'flex'
-});
-
-/**
- * @private
- */
-export const iconWrapperStyle = mergeStyles({
-  pointerEvents: 'none'
-});
