+++ packages/react-components/src/components/styles/RTESendBox.styles.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+/**
+ * Add RTESendBox styles here
+ */
