+++ packages/react-components/src/components/styles/SendBox.styles.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { mergeStyles, IStyle, Theme } from '@fluentui/react';
+import { mergeStyles, Theme } from '@fluentui/react';
 
 /**
  * @private
@@ -75,7 +75,7 @@ export const borderAndBoxShadowStyle = (props: {
   theme: Theme;
   hasErrorMessage: boolean;
   disabled: boolean;
-}): IStyle => {
+}): string => {
   const { theme, hasErrorMessage, disabled } = props;
   const borderColor = hasErrorMessage ? theme.semanticColors.errorText : theme.palette.neutralSecondary;
   const borderColorActive = hasErrorMessage ? theme.semanticColors.errorText : theme.palette.themePrimary;
@@ -83,7 +83,7 @@ export const borderAndBoxShadowStyle = (props: {
   const borderThickness = disabled ? 0 : defaultSendBoxInactiveBorderThicknessREM;
   const borderActiveThickness = disabled ? 0 : defaultSendBoxActiveBorderThicknessREM;
 
-  return {
+  return mergeStyles({
     borderRadius: theme.effects.roundedCorner4,
     border: `${borderThickness}rem solid ${borderColor}`,
 
@@ -95,5 +95,5 @@ export const borderAndBoxShadowStyle = (props: {
       border: `${borderActiveThickness}rem solid ${borderColorActive}`,
       margin: `${defaultSendBoxActiveBorderThicknessREM - borderActiveThickness}rem`
     }
-  };
+  });
 };
