+++ packages/react-components/src/components/utils/SendBoxUtils.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+// Add shared code from Sendbox and RTESendbox
