+++ packages/storybook/stories/INTERNAL/RTE/RTESendBox.stories.tsx
@@ -6,8 +6,8 @@ import { Meta } from '@storybook/react/types-6-0';
 import React from 'react';
 import { COMPONENT_FOLDER_PREFIX } from '../../constants';
 
-const RTESendBoxStory = (args): JSX.Element => {
-  return <RTESendBoxComponent valueToDisplay={args.valueToDisplay} />;
+const RTESendBoxStory = (): JSX.Element => {
+  return <RTESendBoxComponent />;
 };
 
 // This must be the only named export from this module, and must be named to match the storybook path suffix.
@@ -18,8 +18,5 @@ export default {
   id: `${COMPONENT_FOLDER_PREFIX}-internal-rtesendbox`,
   title: `${COMPONENT_FOLDER_PREFIX}/Internal/RTESendBox`,
   component: RTESendBoxComponent,
-  argTypes: {
-    // just a value to be displayed for now but it should be deleted when the component development starts
-    valueToDisplay: { control: 'text', defaultValue: 'Hello World!' }
-  }
+  argTypes: {}
 } as Meta;
